requirements.txt
langchain
langgraph
chromadb
faiss-cpu
pymupdf
streamlit
fpdf2
python-dotenv
